+++ NEWS.md
@@ -7,6 +7,7 @@ Bugs:
 * Major refactor of `rowspan` for HTML files. Thanks to @J-Moravec and @ASKurz for reports #355 and #457.
 * `theme_tt(x, "multipage")` failed to insert `rowhead` in LaTeX. Thanks to @sTeADone for report #460.
 * Error when `x` is a data.table without columns.
+* Adjust Markdown table width when `group_tt(i = )` labels are long. Thanks to @JohannesNE for report #469.
 
 Misc: